remove old samples and keep only one sample file for external demos
I'm creating a new naming system so that our internal samples still get synced through GitHub in the optimize branch, but only the demo sample files are pushed to the main branch for external demos. This way, we can demo the project but not expose our internal dataset while developing the script & conducting our own analyses.